More implementation towards muliplayer client
Adds BaseGameState, LocalGameState, and RemoteGameState classes.

GameState classes accept diff objects and apply mutations to the board
based on those diffs.

The local game state only needs to keep track of one board at a time or
a stack of boards if it supports undo/redo.

The remote game state needs to keep track of the server's board state,
the client's board state, and any diffs that we've sent to the server
but that the server hasn't acknowledged yet. Unacknowledged diffs are
applied locally on top of the last-known server board state so the
changes show up immediately.

In the future, I want to have the ability to apply groups of board diffs
at the same time, so we can implement higher-order operations like
`automaticallyRemoveHints` without server support. BaseGameState could
provide functions implementing said higher-order operations, while board
diffs remain a low-level implementation detail.

I also made some changes to the API that the server provides to better
match the one already implemented in `Gamestate.js`. Those changes are
reflected in the type declarations.