[server] Handle lagged clients with a FullUpdate
Because we don't want the broadcast queue to grow forever, if a client
gets too far behind the broadcast queue, we'll get
`broadcast::RecvError::Lagged`. At this point, we can't trust the diff
broadcast queue and need to catch up:

- Send the client the full board state with a `FullUpdate` message.
- Create a new broadcast receiver, dropping any diffs that might still
  be in the old broadcast receiver's queue.
- Use the last `sync_id` that we saw in the session's `request_receiver`
  task, not the `sync_id` that we got from the broadcast queue, since
  that broadcast queue is no longer valid.

I don't have a good way of testing this yet.