[server] Clean up error handling logic
I was returning `Err(&'static str)` values all over the place, which is
an antipattern (but easier when hacking on something).

This defines a `SudokuError` enum that contains all possible errors that
can be returned to the client.

A nice side-effect of this is that we're able to keep serde_json's error
messages, which (because serde knows about the schema) are very useful
and detailed.