[server] Add room persistence via sqlite
When stopping the server, data is saved to a sqlite database, and when
starting it again, that data is (lazily) read back from sqlite. This
allows you to restart the server without losing data.

Eventually, we should periodically flush data back to disk in case the
server doesn't exit cleanly, and we should evict old inactive rooms from
memory, so we don't grow the heap forever.